agent/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentry-agent"
version = "0.1.0"
description = "Sentry edge agent — RTSP/ONVIF bridge to cloud."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Proprietary" }
authors = [{ name = "Chipmo / Sentry" }]

dependencies = [
    # Vision — headless to keep the image small (no Qt, no GTK).
    "opencv-python-headless==4.10.0.84",
    "numpy>=1.26,<3.0",
    # ONVIF camera discovery + SOAP control.
    "onvif-zeep==0.2.12",
    # HTTP client for cloud API calls.
    "httpx==0.27.2",
    # YAML config.
    "PyYAML==6.0.2",
]

[project.scripts]
sentry-agent = "sentry_agent.cli:main"

[tool.setuptools.packages.find]
include = ["sentry_agent*"]
